Move methods for checking browser logs to TestBenchHelpers
