+++ flow-components-parent/flow-component-demo-helpers/src/main/java/com/vaadin/flow/demo/TabbedComponentDemoTest.java
@@ -18,7 +18,7 @@ package com.vaadin.flow.demo;
 import com.vaadin.testbench.By;
 
 /**
- * Base class for the integration tests for views that use tabs.
+ * Base class for the component integration tests for views that use tabs.
  *
  */
 public abstract class TabbedComponentDemoTest extends ComponentDemoTest {
